Changed Naming from ClientModel to ModelTemplate.
I noticed a potential confusion between flwr clients and ClientModels.
Therefore, I renamed ClientModels to ModelTemplates which is a more honest
description of the responsibility the class undertakes.

Furthermore, I slightly changed the interface of the class since I noticed
common behaviour dependent on different state which I could instead incorporate
into the constructor (to be passed as argument).