Created the base definitions for the Flwr Client.
Flwr needs a client class which implements the functions that will
get executed on the client processes. Therefore, we need to extend
the flwr.client.NumpyClient class and implement all remaining
interfaces in a suitable object.

We combine this client class with our definition of ClientDatasets and
ModelTemplates which get passed by dependency injection.

We have also declared a flwr_parameters folder where we will store
more permanent configurations.