Added Factory Methods to create Datasets with a common API.
In an experiment, we need to create similar datasets where just the
client_identifier (the number identifying the user) changes. Therefore,
we create a factory class interface which defines this behaviour for our
use-cases.

Motivation:
Since the user selection happens within the simulation framework,
we want a way to initialise datasets upon receiving a client number.
The problem is that the dataset class is dependent upon the client
identifier. Therefore, we cannot pass a specific dataset per dependency
injection (since we do not know the datasets we will need on simulation
startup).

Therefore, I wanted a way to pass the configuration parameters of the
datasets as an encapsuled argument which allows us to get user dependent
datasets from an integer.