Increased Performance by automatically deallocating datasets after use
The experiments are very memory intensive, therefore I needed to reduce the amount of data which is in memory at a time.
I did this by changing the ClientDataset class, which now uses context management to deallocate memory once the client
no longer needs it.

Furthermore, I changed the Readme to make note of the heavy memory requirements which suggest at least 30GB of RAM and about
80GB of virtual memory.