Removed Flags Enforcing Determinism
Initial Experiments showed bad convergence of the global model due to the frequent setting of determinacy. This intuitively makes sense since random changes were aligned by determinism causing the model to convergence very early. We potentially want this behaviour in the future, but at the moment it is too unpredictable.